fixed doubling events
I'm registering them when the server starts. However, if the player leaves a world and joins again without closing the game, this counts as the singleplayer server loading again. The event listeners get attached each time the server starts, so now we have a simple latch that says we already loaded them and don't do it again.